Raising an IOError instead of logging an error when fixture file not found
Logging an error when a fixture file couldn't be found and continuing
on was making it hard to debug errors in test scripts, since the
logged error was getting lost in all the output of the failed
tests. Now I simply raise an exception and let the user either remove
the dependency on the fixture or fix it.