Fix for Issue #10
When using JSON as the data serialization method, datetime objects are
not being properly parsed. To fix this, I added a custom decoder that
will use the dateutil.parser.parse function if the dateutil library is
installed, otherwise, it will try to parse the date using
datetime.datetime.strptime and simple format ('%Y-%m-%d'). I also
added some helper functions for printing out an informative message to
the user advising them to install dateutil if it's not already
installed.