Fix for Issue #22
When using the Flask-SQLAlchemy's `init_app` function to initialize
the db object, the app is not added to the current context. This
change checks the both the app and request stacks (depending on the
version of Flask) and makes sure that a request context exists. If one
does not, it creates a new test request context. This should allow
tests to work when `init_app` is used and when the user specifies
their own app and request contexts.